more updates to sln_animal

+++ schemas/+sln_animal/GeneLocus.m
@@ -6,7 +6,7 @@ locus_name              : varchar(16)
 description             : varchar(128)
 
 chromosome = NULL       : tinyint unsigned
-position = NULL         : decimal(2,2) # position along chromosome in cM
+position = NULL         : float # position along chromosome in cM
 
 %}
 classdef GeneLocus < dj.Manual
